Simplify components
