+++ packages/builder/src/pages/builder/workspace/[workspaceId]/agent/[agentId]/operation/[operationId]/index.svelte
@@ -1,21 +1,27 @@
 <script lang="ts">
-  import { notifications } from "@budibase/bbui"
+  import { Body, Icon, notifications, Select } from "@budibase/bbui"
   import {
     FeatureFlag,
     ToolExecutionPrincipal,
+    ToolType,
     type AgentOperation,
     type CaretPositionFn,
+    type EscalationRecipient,
     type InsertAtPositionFn,
   } from "@budibase/types"
   import * as routify from "@roxi/routify"
-  import { onDestroy } from "svelte"
   import TopBar from "@/components/common/TopBar.svelte"
+  import ConfirmDialog from "@/components/common/ConfirmDialog.svelte"
+  import CodeEditor from "@/components/common/CodeEditor/CodeEditor.svelte"
   import {
     EditorModes,
     bindingsToCompletions,
     hbAutocomplete,
   } from "@/components/common/CodeEditor"
+  import EscalationRecipients from "@/components/common/EscalationRecipients.svelte"
+  import LiveToggleButton from "@/components/common/LiveToggleButton.svelte"
   import {
+    contextMenuStore,
     datasources,
     restTemplates,
     workspaceDeploymentStore,
@@ -26,10 +32,19 @@
     featureFlags,
     selectedAgent,
   } from "@/stores/portal"
+  import { bb } from "@/stores/bb"
+  import AgentTabList from "../../AgentTabList.svelte"
+  import AgentUnpublishedChangesIndicator from "../../AgentUnpublishedChangesIndicator.svelte"
+  import GenerateInstructionsControl from "../../GenerateInstructionsControl.svelte"
+  import Knowledge from "../../knowledge/index.svelte"
+  import OperationRailSectionHeader from "../../OperationRailSectionHeader.svelte"
+  import ToolIcon from "../../ToolIcon.svelte"
+  import ToolsDropdown from "../../ToolsDropdown.svelte"
   import WebSearchConfigModal from "../../WebSearchConfigModal.svelte"
   import {
     buildBindingIcons,
     buildReadableToRuntimeBinding,
+    formatAgentToolLabel,
     getAgentWebSearchConfig,
     isWebSearchConfigured,
     resolveAvailableAgentTools,
@@ -40,26 +55,12 @@
     getIncludedToolRuntimeBindings,
   } from "../../toolBindingUtils"
   import type { AgentTool } from "../../toolTypes"
-  import { createSaveCoordinator } from "../../operationSaveCoordinator"
-  import OperationInstructionsPane from "./OperationInstructionsPane.svelte"
-  import OperationSettingsRail from "./OperationSettingsRail.svelte"
-  import {
-    insertOperationToolBinding,
-    removeOperationToolBinding,
-  } from "./operationEditorUtils"
-  import {
-    filterOperationTools,
-    groupToolsBySection,
-  } from "./operationToolFilters"
-  import {
-    hasUnsavedOperationInstructions,
-    mergeResyncedOperation,
-    shouldResyncOperationFromStore,
-  } from "./operationPageUtils"
 
   const { goto, params } = routify
 
   $goto
+
+  type RailTab = "tools" | "knowledge" | "approvals"
 
   let togglingLive = $state(false)
   let saving = $state(false)
@@ -67,9 +68,15 @@
   let lastSavedInstructions = $state("")
   let syncedAgentRev: string | undefined = $state()
   let toolSearch = $state("")
+  let activeTab = $state<RailTab>("tools")
   let insertAtPos: InsertAtPositionFn | undefined = $state()
   let getCaretPosition: CaretPositionFn | undefined = $state()
   let webSearchConfigModal: WebSearchConfigModal | undefined = $state()
+  let removeToolDialog: ConfirmDialog | undefined = $state()
+  let toolToRemove: AgentTool | undefined = $state()
+
+  let flushPromise: Promise<boolean> | null = null
+  let saveAgain = false
 
   let agent = $derived($selectedAgent)
   let agentId = $derived($params.agentId || agent?._id)
@@ -122,19 +129,54 @@
       includedRuntimeBindings.includes(tool.runtimeBinding)
     )
   )
-  let filteredTools = $derived(
-    filterOperationTools({
-      tools: availableTools,
-      toolSearch,
-      escalationEnabled: $featureFlags[FeatureFlag.ESCALATION],
+  let filteredTools = $derived.by(() =>
+    availableTools.filter(tool => {
+      if (
+        tool.sourceType === ToolType.ESCALATION &&
+        !$featureFlags[FeatureFlag.ESCALATION]
+      ) {
+        return false
+      }
+      const query = toolSearch.trim().toLowerCase()
+      return (
+        !query ||
+        `${tool.sourceLabel || ""} ${tool.readableName || tool.name}`
+          .toLowerCase()
+          .includes(query)
+      )
     })
   )
-  let toolSections = $derived(groupToolsBySection(filteredTools))
+  let toolSections = $derived.by(() =>
+    filteredTools.reduce(
+      (sections, tool) => {
+        const section = tool.sourceLabel || "Tools"
+        sections[section] ||= []
+        sections[section].push(tool)
+        return sections
+      },
+      {} as Record<string, AgentTool[]>
+    )
+  )
+
+  const executionPrincipalOptions = [
+    {
+      label: "Requester",
+      value: ToolExecutionPrincipal.REQUESTER,
+    },
+    {
+      label: "Admin (elevated)",
+      value: ToolExecutionPrincipal.ADMIN,
+    },
+  ]
 
   const close = () => $goto("../../config")
 
-  const loadOperation = () => {
+  const hasUnsavedInstructions = () =>
+    operation && (operation.promptInstructions || "") !== lastSavedInstructions
+
+  $effect(() => {
     if (!agent?._id || !operationId) {
+      operation = undefined
       syncedAgentRev = undefined
       return
     }
@@ -145,7 +187,7 @@
       return
     }
 
-    if (operation?.id !== selected.id) {
+    if (operation?.id !== operationId) {
       operation = { ...selected }
       lastSavedInstructions = operation.promptInstructions || ""
       syncedAgentRev = agent._rev
@@ -153,29 +195,16 @@
     }
 
     if (
-      shouldResyncOperationFromStore({
-        agentRev: agent._rev,
-        syncedAgentRev,
-        isSaving: saving || togglingLive,
-      })
+      agent._rev !== syncedAgentRev &&
+      !saving &&
+      !togglingLive &&
+      !hasUnsavedInstructions()
     ) {
-      const preserveInstructionEdits = hasUnsavedOperationInstructions({
-        promptInstructions: operation?.promptInstructions,
-        lastSavedInstructions,
-      })
-      operation = mergeResyncedOperation({
-        storeOperation: selected,
-        localOperation: operation || selected,
-        preserveInstructionEdits,
-      })
-      if (!preserveInstructionEdits) {
-        lastSavedInstructions = operation.promptInstructions || ""
-      }
+      operation = { ...selected }
+      lastSavedInstructions = operation.promptInstructions || ""
       syncedAgentRev = agent._rev
     }
-  }
-
-  $effect(loadOperation)
+  })
 
   const persistOperation = async (): Promise<boolean> => {
     if (!agentId || !operation) {
@@ -219,16 +248,29 @@
     }
   }
 
-  const operationSaveCoordinator = createSaveCoordinator(persistOperation)
-
-  const saveOperation = async (updates: Partial<AgentOperation> = {}) => {
+  const saveOperation = async (
+    updates: Partial<AgentOperation> = {}
+  ): Promise<boolean> => {
     if (!agentId || !operation) {
       return false
     }
     if (Object.keys(updates).length > 0) {
       operation = { ...operation, ...updates }
     }
-    return operationSaveCoordinator.save()
+
+    saveAgain = true
+    if (!flushPromise) {
+      flushPromise = Promise.resolve().then(async () => {
+        let result = false
+        while (saveAgain) {
+          saveAgain = false
+          result = await persistOperation()
+        }
+        flushPromise = null
+        return result
+      })
+    }
+    return flushPromise
   }
 
   const updateInstructions = (instructions: string) => {
@@ -239,14 +281,23 @@
   }
 
   const insertTool = (tool: AgentTool) => {
-    if (!operation) {
+    if (!operation || !tool.readableBinding) {
       return
     }
-    const nextInstructions = insertOperationToolBinding({
-      tool,
-      instructions: operation.promptInstructions || "",
-      getCaretPosition,
-      insertAtPos,
+    const current = operation.promptInstructions || ""
+    const caret = getCaretPosition?.() || {
+      start: current.length,
+      end: current.length,
+    }
+    const binding = `{{ ${tool.readableBinding} }}`
+    const nextInstructions =
+      current.slice(0, caret.start) + binding + current.slice(caret.end)
+
+    insertAtPos?.({
+      start: caret.start,
+      end: caret.end,
+      value: binding,
+      cursor: { anchor: caret.start + binding.length },
     })
     saveOperation({ promptInstructions: nextInstructions })
   }
@@ -255,10 +306,10 @@
     if (!operation || !tool.readableBinding) {
       return
     }
-    const nextInstructions = removeOperationToolBinding({
-      instructions: operation.promptInstructions || "",
-      readableBinding: tool.readableBinding,
-    })
+    const escaped = tool.readableBinding.replace(/[.*+?^${}()|[\]\\]/g, "\\$&")
+    const nextInstructions = (operation.promptInstructions || "")
+      .replace(new RegExp(`\\{\\{\\s*${escaped}\\s*\\}\\}`, "g"), "")
+      .replace(/\n{3,}/g, "\n\n")
     saveOperation({ promptInstructions: nextInstructions })
   }
 
@@ -302,26 +353,52 @@
     }
   }
 
-  const updateRecipients = (recipients: unknown[]) => {
+  const updateRecipients = (
+    recipients: { type: string; config: Record<string, unknown> }[]
+  ) => {
     if (!operation) {
       return
     }
-    operation.escalation = { ...(operation.escalation || {}), recipients }
+    operation.escalation = {
+      ...(operation.escalation || {}),
+      recipients: recipients as EscalationRecipient[],
+    }
     saveOperation()
   }
 
-  const openWebSearchConfigModal = () => {
-    webSearchConfigModal?.show()
+  const confirmRemoveTool = (tool: AgentTool) => {
+    toolToRemove = tool
+    removeToolDialog?.show()
   }
 
-  onDestroy(() => {
-    if (
-      operation &&
-      (operation.promptInstructions || "") !== lastSavedInstructions
-    ) {
-      saveOperation()
+  const handleRemoveToolConfirm = () => {
+    if (!toolToRemove) {
+      return
     }
-  })
+    removeTool(toolToRemove)
+    toolToRemove = undefined
+  }
+
+  const clearToolToRemove = () => {
+    toolToRemove = undefined
+  }
+
+  const openToolMenu = (event: MouseEvent, tool: AgentTool) => {
+    event.preventDefault()
+    event.stopPropagation()
+    contextMenuStore.open(
+      "agent-operation-tool",
+      [
+        {
+          icon: "trash",
+          name: "Remove tool",
+          visible: true,
+          callback: () => confirmRemoveTool(tool),
+        },
+      ],
+      { x: event.clientX, y: event.clientY }
+    )
+  }
 </script>
 
 {#if operation && agentId}
@@ -336,46 +413,204 @@
     />
 
     <div class="operation-content">
-      <OperationInstructionsPane
-        {operation}
-        {toolsLoaded}
-        {promptBindings}
-        {bindingIcons}
-        {completions}
-        {filteredTools}
-        {toolSections}
-        bind:toolSearch
-        bind:insertAtPos
-        bind:getCaretPosition
-        {webSearchConfigured}
-        {togglingLive}
-        onInstructionsChange={updateInstructions}
-        onInstructionsBlur={() => saveOperation()}
-        onApplyGeneratedInstructions={instructions =>
-          saveOperation({ promptInstructions: instructions })}
-        onToggleLive={toggleOperationLive}
-        onToolClick={insertTool}
-        onConfigureWebSearch={openWebSearchConfigModal}
-      />
+      <main class="instructions-pane">
+        <div class="instructions-header">
+          <Body size="S" weight="500">Operation instructions</Body>
+          <div class="instructions-actions">
+            <AgentUnpublishedChangesIndicator />
+            <GenerateInstructionsControl
+              triggerLabel="Help write instructions"
+              promptInstructions={operation.promptInstructions || ""}
+              {promptBindings}
+              {bindingIcons}
+              onApplyInstructions={instructions =>
+                saveOperation({ promptInstructions: instructions })}
+            />
+            <LiveToggleButton
+              live={operation.live === true}
+              size="S"
+              disabled={togglingLive}
+              on:click={toggleOperationLive}
+            />
+          </div>
+        </div>
 
-      <OperationSettingsRail
-        bind:operation
-        {agentId}
-        {includedTools}
-        {filteredTools}
-        {toolSections}
-        bind:toolSearch
-        {webSearchConfigured}
-        onToolClick={insertTool}
-        onConfigureWebSearch={openWebSearchConfigModal}
-        onRemoveTool={removeTool}
-        onSetToolPrincipal={setToolPrincipal}
-        {getEffectiveToolPrincipal}
-        onUpdated={() => saveOperation()}
-        onRecipientsChange={updateRecipients}
-      />
+        <div class="editor-shell">
+          <div class="editor-body">
+            {#if toolsLoaded}
+              <CodeEditor
+                value={operation.promptInstructions || ""}
+                bindings={promptBindings}
+                {bindingIcons}
+                {completions}
+                mode={EditorModes.Handlebars}
+                renderBindingsAsTags
+                renderMarkdownDecorations
+                bind:insertAtPos
+                bind:getCaretPosition
+                on:change={event => updateInstructions(event.detail || "")}
+                on:blur={() => saveOperation()}
+              />
+            {/if}
+          </div>
+          <div class="editor-footer">
+            <span
+              >Use <code>{`{{`}</code> to add tools to your instructions.</span
+            >
+            <div class="tools-popover-container">
+              <ToolsDropdown
+                {filteredTools}
+                {toolSections}
+                bind:toolSearch
+                webSearchEnabled={webSearchConfigured}
+                onToolClick={insertTool}
+                onAddApiConnection={() => bb.settings("/connections/apis")}
+                onConfigureWebSearch={() => webSearchConfigModal?.show()}
+              />
+            </div>
+          </div>
+        </div>
+      </main>
+
+      <aside class="settings-rail">
+        <AgentTabList ariaLabel="Operation settings" bordered>
+          <button
+            class:active={activeTab === "tools"}
+            onclick={() => (activeTab = "tools")}>Tools</button
+          >
+          <button
+            class:active={activeTab === "knowledge"}
+            onclick={() => (activeTab = "knowledge")}>Knowledge</button
+          >
+          {#if $featureFlags[FeatureFlag.ESCALATION]}
+            <button
+              class:active={activeTab === "approvals"}
+              onclick={() => (activeTab = "approvals")}>Approvals</button
+            >
+          {/if}
+        </AgentTabList>
+
+        <div class="rail-content">
+          {#if activeTab === "tools"}
+            <div class="rail-section">
+              <OperationRailSectionHeader
+                title="Tools"
+                description="Give the operation access to the tools it needs to complete requests and take action."
+              >
+                {#snippet actions()}
+                  <div class="tools-popover-container">
+                    <ToolsDropdown
+                      {filteredTools}
+                      {toolSections}
+                      bind:toolSearch
+                      webSearchEnabled={webSearchConfigured}
+                      onToolClick={insertTool}
+                      onAddApiConnection={() =>
+                        bb.settings("/connections/apis")}
+                      onConfigureWebSearch={() => webSearchConfigModal?.show()}
+                    />
+                  </div>
+                {/snippet}
+              </OperationRailSectionHeader>
+              <div class="tools-list" role="list">
+                {#each includedTools as tool (tool.runtimeBinding)}
+                  <div
+                    class="tool-row"
+                    class:tool-row--with-run-as={$featureFlags[
+                      FeatureFlag.AI_AGENT_TOOL_SECURITY
+                    ] && tool.executionPolicy.mode === "configurable"}
+                    role="listitem"
+                    oncontextmenu={event => openToolMenu(event, tool)}
+                  >
+                    <div class="tool-row-main">
+                      <div class="tool-name">
+                        <span class="tool-icon">
+                          <ToolIcon
+                            icon={tool.icon}
+                            size="S"
+                            fallbackIcon="Wrench"
+                          />
+                        </span>
+                        <span>{tool.readableBinding}</span>
+                      </div>
+                      <button
+                        aria-label={`Actions for ${tool.readableBinding}`}
+                        onclick={event => openToolMenu(event, tool)}
+                      >
+                        <Icon name="dots-three" size="XS" />
+                      </button>
+                    </div>
+                    {#if $featureFlags[FeatureFlag.AI_AGENT_TOOL_SECURITY] && tool.executionPolicy.mode === "configurable"}
+                      <div
+                        class="tool-row-run-as"
+                        role="group"
+                        aria-label="Run as"
+                        oncontextmenu={event => event.stopPropagation()}
+                      >
+                        <span class="run-as-label">Run as</span>
+                        <Select
+                          size="S"
+                          bordered={false}
+                          placeholder={false}
+                          autoWidth
+                          popoverAutoWidth
+                          value={getEffectiveToolPrincipal(tool)}
+                          options={executionPrincipalOptions}
+                          getOptionLabel={option => option.label}
+                          getOptionValue={option => option.value}
+                          tooltip={`Execution identity for ${formatAgentToolLabel(tool)}`}
+                          on:change={event =>
+                            setToolPrincipal({
+                              toolName: tool.runtimeBinding,
+                              executionPrincipal:
+                                event.detail as ToolExecutionPrincipal,
+                            })}
+                        />
+                      </div>
+                    {/if}
+                  </div>
+                {:else}
+                  <Body size="XS" color="var(--spectrum-global-color-gray-700)"
+                    >No tools are referenced in these instructions.</Body
+                  >
+                {/each}
+              </div>
+            </div>
+          {:else if activeTab === "knowledge"}
+            <Knowledge bind:operation onUpdated={() => saveOperation()} />
+          {:else}
+            <div class="rail-section approval-panel">
+              <OperationRailSectionHeader
+                title="Approvals"
+                description="Choose who gets notified when this operation escalates for approval."
+              />
+              <EscalationRecipients
+                single
+                recipients={operation.escalation?.recipients || []}
+                {agentId}
+                onChange={updateRecipients}
+              />
+            </div>
+          {/if}
+        </div>
+      </aside>
     </div>
   </div>
+
+  <ConfirmDialog
+    bind:this={removeToolDialog}
+    title="Remove tool?"
+    okText="Remove"
+    warning={true}
+    onOk={handleRemoveToolConfirm}
+    onCancel={clearToolToRemove}
+    onClose={clearToolToRemove}
+  >
+    {#if toolToRemove?.readableBinding}
+      Remove <b>{toolToRemove.readableBinding}</b> from this operation? Its binding
+      will also be removed from the instructions.
+    {/if}
+  </ConfirmDialog>
 
   <WebSearchConfigModal
     bind:this={webSearchConfigModal}
@@ -391,12 +626,192 @@
     flex-direction: column;
     background: var(--background);
   }
+
   .operation-content {
     display: grid;
     grid-template-columns: minmax(0, 1fr) 360px;
     flex: 1 1 auto;
     min-height: 0;
   }
+
+  .instructions-pane {
+    display: flex;
+    min-width: 0;
+    min-height: 0;
+    flex-direction: column;
+    gap: 10px;
+    padding: 10px 12px 12px;
+  }
+
+  .instructions-header {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    gap: var(--spacing-m);
+  }
+
+  .instructions-actions {
+    display: flex;
+    align-items: center;
+    gap: var(--spacing-m);
+  }
+
+  .editor-shell {
+    display: flex;
+    min-height: 0;
+    flex: 1 1 auto;
+    flex-direction: column;
+    overflow: hidden;
+    border: 1px solid var(--spectrum-global-color-gray-200);
+    border-radius: 6px;
+    background: var(--spectrum-global-color-gray-100);
+  }
+
+  .editor-body {
+    min-height: 0;
+    flex: 1 1 auto;
+    overflow: auto;
+  }
+
+  .editor-body :global(.cm-editor) {
+    min-height: 100%;
+    height: 100%;
+    background: var(--spectrum-global-color-gray-100) !important;
+  }
+
+  .editor-footer {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    gap: var(--spacing-m);
+    padding: 8px 12px;
+    border-top: 1px solid var(--spectrum-global-color-gray-200);
+    font-size: 12px;
+  }
+
+  .editor-footer code {
+    padding: 2px 5px;
+    border-radius: 3px;
+    background: var(--spectrum-global-color-gray-200);
+  }
+
+  .settings-rail {
+    display: flex;
+    min-width: 0;
+    min-height: 0;
+    flex-direction: column;
+    background: var(--background);
+    border-left: 1px solid var(--spectrum-global-color-gray-200);
+  }
+
+  .rail-content {
+    min-height: 0;
+    flex: 1 1 auto;
+    overflow: auto;
+    padding: 20px 12px;
+  }
+
+  .rail-section {
+    display: flex;
+    flex-direction: column;
+    gap: 16px;
+  }
+
+  .tools-list {
+    display: flex;
+    flex-direction: column;
+    gap: 6px;
+  }
+
+  .tool-row {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    gap: 8px;
+    min-height: 34px;
+    padding: 0 12px;
+    border-radius: 4px;
+    background: var(--background-alt);
+    cursor: pointer;
+  }
+
+  .tool-row--with-run-as {
+    flex-direction: column;
+    align-items: stretch;
+    gap: 6px;
+    padding: 8px 12px;
+  }
+
+  .tool-row-main {
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    gap: 8px;
+    min-width: 0;
+    flex: 1 1 auto;
+  }
+
+  .tool-row--with-run-as .tool-row-main {
+    width: 100%;
+  }
+
+  .tool-name {
+    display: flex;
+    min-width: 0;
+    align-items: center;
+    gap: 7px;
+    font-size: 13px;
+    line-height: 17px;
+  }
+
+  .tool-icon {
+    display: flex;
+    width: 14px;
+    height: 14px;
+    flex: 0 0 14px;
+    align-items: center;
+    justify-content: center;
+    overflow: hidden;
+  }
+
+  .tool-icon :global(img),
+  .tool-icon :global(svg) {
+    width: 14px !important;
+    height: 14px !important;
+  }
+
+  .tool-name span {
+    overflow: hidden;
+    text-overflow: ellipsis;
+    white-space: nowrap;
+  }
+
+  .tool-row-run-as {
+    display: flex;
+    align-items: center;
+    gap: var(--spacing-s);
+    padding-left: 21px;
+  }
+
+  .run-as-label {
+    color: var(--spectrum-global-color-gray-700);
+    font-size: var(--font-size-xs);
+    white-space: nowrap;
+  }
+
+  .tool-row-run-as :global(.spectrum-Picker) {
+    min-width: 0;
+  }
+
+  .tool-row-main button {
+    display: flex;
+    border: 0;
+    padding: 4px;
+    background: transparent;
+    color: inherit;
+    cursor: pointer;
+  }
+
   @media (max-width: 900px) {
     .operation-content {
       grid-template-columns: minmax(0, 1fr) 300px;
